third_party/blink/web_tests/external/wpt/css/css-overflow/scroll-markers/scroll-marker-navigation-cycles-expected.txt

This is a testharness.js-based test.
[FAIL] CSS Test: ::scroll-marker key navigation cycles
  assert_equals: last ::scroll-marker gets activated by going from the first one expected "rgb(0, 128, 0)" but got "rgb(0, 0, 255)"
Harness: the test ran to completion.
